added a goalBias factor to RRT::Tree
Previously, all tree growth was random.  There is now an option to bias growth towards
the goal state.  Note that adding this feature reduced the generality of this RRT implementation a bit.
The goal state used to not be specified directly, but was specified by providing a callback that returned
a boolean indicating whether a given state was at the goal or not.  This allowed the RRT::Tree to be used
in cases where the goal was dynamic, rather than being a fixed state.  Although this generality can be
useful for some, I decided to remove it here to enable the RRT to do more for our applications.